Small Fix in Verify Lidar, Change angeToApproxSteps ---------- - fixed a small bug in the verify liar stage that wa causing display issues with the reading from the liar - changed the function angleToApproxSteps, simplified it more and uses lroundf function to round the   number of calculated steps up or down instead of truncating the decimal value

+++ main_mob_platform/lib/Lidar360/Lidar360.h
@@ -9,6 +9,7 @@
 
 #include <Arduino.h>
 #include <I2C.h>
+#include <math.h>
 #include <AccelStepper.h>
 #include <LiquidCrystal.h>
 #include "PinMap.h"
@@ -55,7 +56,7 @@ class Lidar360
         int     llGetDistance();
         int     llGetDistanceAverage(int numberOfReadings);
 
-        int     angleToApproxSteps(int angle);
+        long     angleToApproxSteps(int angle);
 
         int     _buttonA;
         int     _buttonB;
